Removing 't' case for SIXEL patch ref. #35.
	#if SIXEL_PATCH
	case 't':
		/* TODO should probably not be hard-coded */
		ttywrite(";420;720t", 10, 1);
		break;
	#endif // SIXEL_PATCH

This would result in printing ";420;720t" when exiting neovim.

Without this code a line is written to standard err instead:

erresc: unknown csi ESC[23;0t

The ttywrite was added as part of this commit:
   - https://github.com/charlesdaniels/st/commit/b50be8225dd2cd6de35a9b6219c556bac355255c

which states:

> When a S or T CSI escape was encountered, the lines which were scrolled
> away would be deleted from the scrollback buffer. This has been
> corrected - the lines are now preseved.
>
> This fixes a bug where issuing `clear` followed by `lsix` would cause
> the line on which the `lsix` was issued to disappear from the scrollback
> buffer.
>
> Note that the line may scroll out of view and thus dissapear, but it
> will now be preserved in the scrollback buffer.

Given that we could not reproduce the above bug without the ttywrite in
this case I am not convinced that this is actually needed. Leaving this
here in case this comes up again in the future.